+++ relay-server/src/services/store.rs
@@ -39,6 +39,7 @@ use relay_threading::AsyncPool;
 
 use crate::envelope::{AttachmentType, Envelope, Item, ItemType};
 use crate::metrics::{ArrayEncoding, BucketEncoder, MetricOutcomes};
+use crate::processing::{Counted, Managed, OutcomeError, Quantities};
 use crate::service::ServiceError;
 use crate::services::global_config::GlobalConfigHandle;
 use crate::services::outcome::{DiscardItemType, DiscardReason, Outcome, TrackOutcome};
@@ -57,6 +58,14 @@ pub enum StoreError {
     EncodingFailed(std::io::Error),
     #[error("failed to store event because event id was missing")]
     NoEventId,
+}
+
+impl OutcomeError for StoreError {
+    type Error = Self;
+
+    fn consume(self) -> (Option<Outcome>, Self::Error) {
+        (Some(Outcome::Invalid(DiscardReason::Internal)), self)
+    }
 }
 
 struct Producer {
@@ -98,14 +107,42 @@ pub struct StoreMetrics {
     pub retention: u16,
 }
 
+/// Publishes a log item to Sentry core application through Kafka.
+#[derive(Debug)]
+pub struct StoreLog {
+    /// The final trace item which will be produced to Kafka.
+    pub trace_item: TraceItem,
+    /// Outcomes to be emitted when successfully producing the item to Kafka.
+    ///
+    /// Note: this is only a temporary measure, long term these outcomes will be part of the trace
+    /// item and emitted by Snuba to guarantee a delivery to storage.
+    pub quantities: Quantities,
+}
+
+impl Counted for StoreLog {
+    fn quantities(&self) -> Quantities {
+        self.quantities.clone()
+    }
+}
+
 /// The asynchronous thread pool used for scheduling storing tasks in the envelope store.
 pub type StoreServicePool = AsyncPool<BoxFuture<'static, ()>>;
 
 /// Service interface for the [`StoreEnvelope`] message.
 #[derive(Debug)]
 pub enum Store {
+    /// An envelope containing a mixture of items.
+    ///
+    /// Note: Some envelope items are not supported to be submitted at all or through an envelope,
+    /// for example logs must be submitted via [`Self::Log`] instead.
+    ///
+    /// Long term this variant is going to be replaced with fully typed variants of items which can
+    /// be stored instead.
     Envelope(StoreEnvelope),
+    /// Aggregated generic metrics.
     Metrics(StoreMetrics),
+    /// A singular log item.
+    Log(Managed<StoreLog>),
 }
 
 impl Store {
@@ -114,6 +151,7 @@ impl Store {
         match self {
             Store::Envelope(_) => "envelope",
             Store::Metrics(_) => "metrics",
+            Store::Log(_) => "log",
         }
     }
 }
@@ -133,6 +171,14 @@ impl FromMessage<StoreMetrics> for Store {
 
     fn from_message(message: StoreMetrics, _: ()) -> Self {
         Self::Metrics(message)
+    }
+}
+
+impl FromMessage<Managed<StoreLog>> for Store {
+    type Response = NoResponse;
+
+    fn from_message(message: Managed<StoreLog>, _: ()) -> Self {
+        Self::Log(message)
     }
 }
 
@@ -171,6 +217,7 @@ impl StoreService {
             match message {
                 Store::Envelope(message) => self.handle_store_envelope(message),
                 Store::Metrics(message) => self.handle_store_metrics(message),
+                Store::Log(message) => self.handle_store_log(message),
             }
         })
     }
@@ -296,7 +343,6 @@ impl StoreService {
                 ItemType::Span => {
                     self.produce_span(scoping, received_at, event_id, retention, item)?
                 }
-                ItemType::Log => self.produce_log(scoping, received_at, retention, item)?,
                 ItemType::ProfileChunk => self.produce_profile_chunk(
                     scoping.organization_id,
                     scoping.project_id,
@@ -304,6 +350,18 @@ impl StoreService {
                     retention,
                     item,
                 )?,
+                // Explicitly unsupported items which must be submitted via a specific store message.
+                ty @ ItemType::Log => {
+                    debug_assert!(
+                        false,
+                        "received {ty} through an envelope, \
+                        this item must be submitted via a specific store message instead"
+                    );
+                    relay_log::error!(
+                        tags.project_key = %scoping.project_key,
+                        "StoreService received unsupported item type '{ty}' in envelope"
+                    );
+                }
                 other => {
                     let event_type = event_item.as_ref().map(|item| item.ty().as_str());
                     let item_types = envelope
@@ -465,6 +523,46 @@ impl StoreService {
                 gauge(RelayGauges::MetricDelayMax) = max,
                 namespace = namespace.as_str()
             );
+        }
+    }
+
+    fn handle_store_log(&self, message: Managed<StoreLog>) {
+        let scoping = message.scoping();
+        let received_at = message.received_at();
+
+        let quantities = message.try_accept(|log| {
+            let message = KafkaMessage::Item {
+                headers: BTreeMap::from([
+                    ("project_id".to_owned(), scoping.project_id.to_string()),
+                    (
+                        "item_type".to_owned(),
+                        (TraceItemType::Log as i32).to_string(),
+                    ),
+                ]),
+                message: log.trace_item,
+                item_type: TraceItemType::Log,
+            };
+
+            self.produce(KafkaTopic::Items, message)
+                .map(|()| log.quantities)
+        });
+
+        // Accepted outcomes when items have been successfully produced to rdkafka.
+        //
+        // This is only a temporary measure, long term these outcomes will be part of the trace
+        // item and emitted by Snuba to guarantee a delivery to storage.
+        if let Ok(quantities) = quantities {
+            for (category, quantity) in quantities {
+                self.outcome_aggregator.send(TrackOutcome {
+                    category,
+                    event_id: None,
+                    outcome: Outcome::Accepted,
+                    quantity: u32::try_from(quantity).unwrap_or(u32::MAX),
+                    remote_addr: None,
+                    scoping,
+                    timestamp: received_at,
+                });
+            }
         }
     }
 
@@ -1188,141 +1286,6 @@ impl StoreService {
         Ok(())
     }
 
-    fn produce_log(
-        &self,
-        scoping: Scoping,
-        received_at: DateTime<Utc>,
-        retention_days: u16,
-        item: &Item,
-    ) -> Result<(), StoreError> {
-        relay_log::trace!("Producing log");
-
-        let payload = item.payload();
-        let d = &mut Deserializer::from_slice(&payload);
-        let logs: LogKafkaMessages = match serde_path_to_error::deserialize(d) {
-            Ok(logs) => logs,
-            Err(error) => {
-                relay_log::error!(
-                    error = &error as &dyn std::error::Error,
-                    "failed to parse log"
-                );
-                self.outcome_aggregator.send(TrackOutcome {
-                    category: DataCategory::LogItem,
-                    event_id: None,
-                    outcome: Outcome::Invalid(DiscardReason::InvalidLog),
-                    quantity: 1,
-                    remote_addr: None,
-                    scoping,
-                    timestamp: received_at,
-                });
-                self.outcome_aggregator.send(TrackOutcome {
-                    category: DataCategory::LogByte,
-                    event_id: None,
-                    outcome: Outcome::Invalid(DiscardReason::InvalidLog),
-                    quantity: payload.len() as u32,
-                    remote_addr: None,
-                    scoping,
-                    timestamp: received_at,
-                });
-                return Ok(());
-            }
-        };
-
-        let num_logs = logs.items.len() as u32;
-        for log in logs.items {
-            let timestamp_seconds = log.timestamp as i64;
-            let timestamp_nanos = (log.timestamp.fract() * 1e9) as u32;
-            let item_id = u128::from_be_bytes(
-                *Uuid::new_v7(uuid::Timestamp::from_unix(
-                    uuid::NoContext,
-                    timestamp_seconds as u64,
-                    timestamp_nanos,
-                ))
-                .as_bytes(),
-            )
-            .to_le_bytes()
-            .to_vec();
-            let mut trace_item = TraceItem {
-                item_type: TraceItemType::Log.into(),
-                organization_id: scoping.organization_id.value(),
-                project_id: scoping.project_id.value(),
-                received: Some(Timestamp {
-                    seconds: safe_timestamp(received_at) as i64,
-                    nanos: 0,
-                }),
-                retention_days: retention_days.into(),
-                timestamp: Some(Timestamp {
-                    seconds: timestamp_seconds,
-                    nanos: 0,
-                }),
-                trace_id: log.trace_id.to_string(),
-                item_id,
-                attributes: Default::default(),
-                client_sample_rate: 1.0,
-                server_sample_rate: 1.0,
-            };
-
-            for (name, attribute) in log.attributes.unwrap_or_default() {
-                if let Some(attribute_value) = attribute {
-                    if let Some(v) = attribute_value.value {
-                        let any_value = match v {
-                            LogAttributeValue::String(value) => AnyValue {
-                                value: Some(Value::StringValue(value)),
-                            },
-                            LogAttributeValue::Int(value) => AnyValue {
-                                value: Some(Value::IntValue(value)),
-                            },
-                            LogAttributeValue::Bool(value) => AnyValue {
-                                value: Some(Value::BoolValue(value)),
-                            },
-                            LogAttributeValue::Double(value) => AnyValue {
-                                value: Some(Value::DoubleValue(value)),
-                            },
-                            LogAttributeValue::Unknown(_) => continue,
-                        };
-
-                        trace_item.attributes.insert(name.into(), any_value);
-                    }
-                }
-            }
-
-            let message = KafkaMessage::Log {
-                headers: BTreeMap::from([
-                    ("project_id".to_owned(), scoping.project_id.to_string()),
-                    (
-                        "item_type".to_owned(),
-                        (TraceItemType::Log as i32).to_string(),
-                    ),
-                ]),
-                message: trace_item,
-            };
-
-            self.produce(KafkaTopic::Items, message)?;
-        }
-
-        // We need to track the count and bytes separately for possible rate limits and quotas on both counts and bytes.
-        self.outcome_aggregator.send(TrackOutcome {
-            category: DataCategory::LogItem,
-            event_id: None,
-            outcome: Outcome::Accepted,
-            quantity: num_logs,
-            remote_addr: None,
-            scoping,
-            timestamp: received_at,
-        });
-        self.outcome_aggregator.send(TrackOutcome {
-            category: DataCategory::LogByte,
-            event_id: None,
-            outcome: Outcome::Accepted,
-            quantity: payload.len() as u32,
-            remote_addr: None,
-            scoping,
-            timestamp: received_at,
-        });
-
-        Ok(())
-    }
-
     fn produce_profile_chunk(
         &self,
         organization_id: OrganizationId,
@@ -1800,44 +1763,6 @@ impl SpanKafkaMessage<'_> {
     }
 }
 
-#[derive(Clone, Debug, Deserialize)]
-#[serde(tag = "type", content = "value")]
-enum LogAttributeValue {
-    #[serde(rename = "string")]
-    String(String),
-    #[serde(rename = "boolean")]
-    Bool(bool),
-    #[serde(rename = "integer")]
-    Int(i64),
-    #[serde(rename = "double")]
-    Double(f64),
-    #[serde(rename = "unknown")]
-    Unknown(()),
-}
-
-/// This is a temporary struct to convert the old attribute format to the new one.
-#[derive(Debug, Deserialize)]
-#[allow(dead_code)]
-struct LogAttribute {
-    #[serde(flatten)]
-    value: Option<LogAttributeValue>,
-}
-
-#[derive(Debug, Deserialize)]
-struct LogKafkaMessages<'a> {
-    #[serde(borrow)]
-    items: Vec<LogKafkaMessage<'a>>,
-}
-
-#[derive(Debug, Deserialize)]
-struct LogKafkaMessage<'a> {
-    trace_id: EventId,
-    #[serde(default)]
-    timestamp: f64,
-    #[serde(borrow, default)]
-    attributes: Option<BTreeMap<&'a str, Option<LogAttribute>>>,
-}
-
 fn none_or_empty_object(value: &Option<&RawValue>) -> bool {
     match value {
         None => true,
@@ -1898,11 +1823,6 @@ enum KafkaMessage<'a> {
         #[serde(flatten)]
         message: SpanKafkaMessage<'a>,
     },
-    Log {
-        headers: BTreeMap<String, String>,
-        #[serde(skip)]
-        message: TraceItem,
-    },
     ProfileChunk(ProfileChunkKafkaMessage),
 }
 
@@ -1925,7 +1845,6 @@ impl Message for KafkaMessage<'_> {
             KafkaMessage::ReplayEvent(_) => "replay_event",
             KafkaMessage::ReplayRecordingNotChunked(_) => "replay_recording_not_chunked",
             KafkaMessage::CheckIn(_) => "check_in",
-            KafkaMessage::Log { .. } => "log",
             KafkaMessage::Span { .. } => "span",
             KafkaMessage::ProfileChunk(_) => "profile_chunk",
             KafkaMessage::Item { item_type, .. } => item_type.as_str_name(),
@@ -1958,7 +1877,6 @@ impl Message for KafkaMessage<'_> {
 
             // Random partitioning
             Self::Profile(_)
-            | Self::Log { .. }
             | Self::ReplayRecordingNotChunked(_)
             | Self::ProfileChunk(_)
             | Self::Item { .. } => None,
@@ -1971,8 +1889,7 @@ impl Message for KafkaMessage<'_> {
         match &self {
             KafkaMessage::Metric { headers, .. }
             | KafkaMessage::Span { headers, .. }
-            | KafkaMessage::Item { headers, .. }
-            | KafkaMessage::Log { headers, .. } => {
+            | KafkaMessage::Item { headers, .. } => {
                 if !headers.is_empty() {
                     return Some(headers);
                 }
@@ -2000,13 +1917,11 @@ impl Message for KafkaMessage<'_> {
             KafkaMessage::Span { message, .. } => serde_json::to_vec(message)
                 .map(Cow::Owned)
                 .map_err(ClientError::InvalidJson),
-            KafkaMessage::Log { message, .. } | KafkaMessage::Item { message, .. } => {
+            KafkaMessage::Item { message, .. } => {
                 let mut payload = Vec::new();
-
                 if message.encode(&mut payload).is_err() {
                     return Err(ClientError::ProtobufEncodingFailed);
                 }
-
                 Ok(Cow::Owned(payload))
             }
             _ => rmp_serde::to_vec_named(&self)